pyproject.toml
[project]
name = "pqr"
version = "0.1.0"
description = "Generate QR Codes for 3D Prints"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pillow>=11",
  "pydantic>=2",
  "pyyaml>=6",
  "qrcode>=8",
  "rich>=13",
  "tomli-w>=1",
  "typer>=0.13",
]

[[project.authors]]
name = "Shant Ergenian"
email = "shaaant@pm.me"

[project.urls]
repository = "https://github.com/tnahs/PrintQR"

[project.scripts]
pqr = "pqr:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"pqr.data" = ["*.toml"]
"pqr.resources.fonts" = ["*.ttf"]
